+++ src/main/java/com/gmail/netcracker/application/dto/dao/interfaces/NoteDao.java
@@ -9,7 +9,7 @@ public interface NoteDao {
 
     List<Note> noteList();
 
-    Note getNote(int noteId);
+    Note getNote(Long noteId);
 
     void delete(Long noteId);
