+++ src/main/java/com/gmail/netcracker/application/service/interfaces/NoteService.java
@@ -8,7 +8,7 @@ public interface NoteService {
 
     List<Note> noteList();
 
-    Note getNote(int noteId);
+    Note getNote(Long noteId);
 
     void delete(Long noteId);
